refactor(sdk/reader): move claude parse engine into claude/incremental.rs
claude.rs was the largest SDK source file (2134 lines). Extract the
incremental-parse engine — ClaudeParseState + its handlers, run_incremental,
prescan_nodes, record_root_incremental, PrescanOutput — into a new
claude/incremental.rs submodule, leaving the root as the public API +
parse-primitive helpers. claude.rs drops to 1345; incremental.rs is 844.

Pure code-movement: the moved block is byte-identical (only run_incremental
gained pub(super)); root helpers the engine uses were widened to
pub(in crate::reader::claude) (tightest scope, warning-free). Gives the reader
trio the same shape (each reader is a directory; the engine lives in its own
submodule). 103 claude tests pass, full SDK green, clippy clean, live OK.
